Rename java-runtime-metrics to runtime-metrics
