+++ instrumentation/runtime-metrics/library/src/main/java/io/opentelemetry/instrumentation/runtimemetrics/MemoryPools.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.java.runtime;
+package io.opentelemetry.instrumentation.runtimemetrics;
 
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.common.Labels;
