+++ settings.gradle
@@ -106,7 +106,6 @@ include ':instrumentation:java-classloader:osgi-testing'
 include ':instrumentation:java-classloader:tomcat-testing'
 include ':instrumentation:java-concurrent'
 include ':instrumentation:java-httpclient'
-include ':instrumentation:java-runtime-metrics:library'
 include ':instrumentation:jaxrs:jaxrs-1.0'
 include ':instrumentation:jaxrs:jaxrs-2.0:jaxrs-2.0-common'
 include ':instrumentation:jaxrs:jaxrs-2.0:jaxrs-2.0-jersey-2.0'
@@ -168,6 +167,7 @@ include ':instrumentation:reactor-netty-0.9'
 include ':instrumentation:rediscala-1.8'
 include ':instrumentation:redisson-3.0'
 include ':instrumentation:rmi'
+include ':instrumentation:runtime-metrics:library'
 include ':instrumentation:rxjava-1.0'
 include ':instrumentation:rxjava-1.0'
 include ':instrumentation:scala-executors'
